Add the ability for candore to resume an extraction after an error
This change makes some structural changes to allow the extractor to
store extraction to files that it can later resume from.
Due to the async nature of the extractor, I haven't allowed this
save/resume to happen within an entity, so the latest entity being
extracted would lose its progress and have to start again.